release(all): pbrAudio 0.2.4 beta

+++ src/rigidBody/core/rigidbody_engine.py
@@ -83,7 +83,7 @@ class rigidBodyEngine:
             if os.path.exists(f"{self.trajectories_dir}") and not len(os.listdir(f"{self.trajectories_dir}")) == 0:
 #                trajectories_idx = 0
                 for filename in os.listdir(f"{self.trajectories_dir}"):
-                    if filename.endswith('.pkl'):
+                    if os.path.isfile(filename) and filename.endswith('.pkl'):
                         trajectory = TrajectoryData.load(f"{self.trajectories_dir}/{filename}")
                         _ = self.entity_manager.register('trajectories', trajectory)
 #                        self.entity_manager.register('trajectories', trajectories, trajectories_idx)
@@ -100,7 +100,7 @@ class rigidBodyEngine:
         if len(collisions) == 0:
             if os.path.exists(f"{self.collisions_dir}") and not len(os.listdir(f"{self.collisions_dir}")) == 0:
                 for filename in os.listdir(f"{self.collisions_dir}"):
-                    if filename.endswith('.pkl'):
+                    if os.path.isfile(filename) and filename.endswith('.pkl'):
 #                        idx = int(filename.removesuffix('.pkl'))
                         collisions = CollisionData.load(f"{self.collisions_dir}/{filename}")
 #                        self.entity_manager.register('collisions', collisions, idx)
@@ -111,7 +111,7 @@ class rigidBodyEngine:
             if os.path.exists(f"{self.forces_dir}") and not len(os.listdir(f"{self.forces_dir}")) == 0:
 #                forces_idx = 0
                 for filename in os.listdir(f"{self.forces_dir}"):
-                    if filename.endswith('.pkl'):
+                    if os.path.isfile(filename) and filename.endswith('.pkl'):
                         forces = ForceDataSequence.load(f"{self.forces_dir}/{filename}")
                         _ = self.entity_manager.register('forces', forces)
 #                        self.entity_manager.register('forces', forces, forces_idx)
@@ -124,8 +124,9 @@ class rigidBodyEngine:
                 filenames = os.listdir(self.modalvertices_dir)
 #                modalvertices_idx = 0
                 for filename in filenames:
-                    modal_vertices = ModalVertices.load(f"{self.modalvertices_dir}/{filename}")
-                    _ = self.entity_manager.register('modal_vertices', modal_vertices)
+                    if os.path.isfile(filename):
+                        modal_vertices = ModalVertices.load(f"{self.modalvertices_dir}/{filename}")
+                        _ = self.entity_manager.register('modal_vertices', modal_vertices)
 #                    self.entity_manager.register('modal_vertices', modal_vertices, modalvertices_idx)
 #                    modalvertices_idx += 1
 
@@ -138,10 +139,12 @@ class rigidBodyEngine:
                 filenames = os.listdir(self.scoretracks_dir)
 #                scoretracks_idx = 0
                 for filename in filenames:
-                    score_tracks = ScoreTrack.load(f"{self.scoretracks_dir}/{filename}")
-                    _ = self.entity_manager.register('score_tracks', score_tracks)
+                    if os.path.isfile(filename):
+                        score_tracks = ScoreTrack.load(f"{self.scoretracks_dir}/{filename}")
+                        _ = self.entity_manager.register('score_tracks', score_tracks)
                 for filename in filenames:
-                    os.remove(f"{self.scoretracks_dir}/{filename}")
+                    if os.path.isfile(filename):
+                        os.remove(f"{self.scoretracks_dir}/{filename}")
 #                    self.entity_manager.register('score_tracks', score_tracks, scoretracks_idx)
 #                    scoretracks_idx += 1
 
@@ -192,8 +195,9 @@ class rigidBodyEngine:
                 filenames = os.listdir(self.scoretracks_dir)
 #                scoretracks_idx = 0
                 for filename in filenames:
-                    score_tracks = ScoreTrack.load(f"{self.scoretracks_dir}/{filename}", final=True)
-                    _ = self.entity_manager.register('score_tracks', score_tracks)
+                    if os.path.isfile(filename):
+                        score_tracks = ScoreTrack.load(f"{self.scoretracks_dir}/{filename}", final=True)
+                        _ = self.entity_manager.register('score_tracks', score_tracks)
 #                    self.entity_manager.register('score_tracks', score_tracks, scoretracks_idx)
 #                    scoretracks_idx += 1
